Move file from seperated to one file myGUI.py
