--- pages/prediction_page.py
@@ -1,6 +1,0 @@
-import tkinter as tk
-
-def create_prediction_page(notebook):
-    predict_page = tk.Frame(notebook)
-    notebook.add(predict_page, text="Predict Phase")
-    # Add widgets and functionality for the prediction page as needed
