+++ requirements.txt
@@ -2,15 +2,20 @@ click==8.1.7
 contourpy==1.2.0
 cycler==0.12.1
 fonttools==4.45.0
+igraph==0.11.3
 joblib==1.3.2
 kiwisolver==1.4.5
 matplotlib==3.8.2
+networkx==3.2.1
 nltk==3.8.1
 numpy==1.26.2
 packaging==23.2
 Pillow==10.1.0
+pycairo==1.25.1
 pyparsing==3.1.1
 python-dateutil==2.8.2
 regex==2023.10.3
+scipy==1.11.4
 six==1.16.0
+texttable==1.7.0
 tqdm==4.66.1
